Extract `State` tests to a separate file
